feat: refactor and merge code
