--- src/Lamba-function/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "Lamba-function"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-lambda_runtime = "0.10.0"
-aws_lambda_events = "0.15.0"
-serde = "1"
-serde_json = "1.0"
-tokio = { version = "1", features = ["macros"] }
-aws-config = "1.1.8"
-aws-smithy-http = "0.60.7"
-aws-sdk-s3 = "1.20.0"
-thumbnailer = "0.5.1"
-mime = "0.3.16"
-async-trait = "0.1.68"
-
-
