+++ src/lambda/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "format"
+name = "get_receipt"
 version = "0.1.0"
 edition = "2021"
 
@@ -15,11 +15,14 @@ edition = "2021"
 # and it will keep the alphabetic ordering for you.
 
 [dependencies]
-aws_lambda_events = { version = "0.15.0", default-features = false, features = ["config"] }
-
-lambda_runtime = "0.10.0"
-serde = { version = "1.0.197", features = ["derive"] }
-serde_derive = "1.0.197"
-serde_json = "1.0.114"
+lambda_http = "0.10.0"
 tokio = { version = "1", features = ["macros"] }
-
+serde = "1"
+serde_json = "1.0"
+tokio = { version = "1", features = ["macros"] }
+aws-config = "1.1.8"
+aws-smithy-http = "0.60.7"
+aws-sdk-s3 = "1.20.0"
+thumbnailer = "0.5.1"
+mime = "0.3.16"
+async-trait = "0.1.68"
