+++ src/lambda/src/format/block/response_block.rs
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+
+#[derive(Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct ResponseBlock {
+    confidence: f64,
+    text: String,
+}
