+++ src/lambda/src/format/block/text_type.rs
@@ -1,0 +1,10 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+
+#[derive(Serialize, Deserialize)]
+pub enum TextType {
+    #[serde(rename = "PRINTED")]
+    Printed,
+}
