+++ src/lambda/src/format/block/type_enum.rs
@@ -1,0 +1,10 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+
+#[derive(Serialize, Deserialize)]
+pub enum TypeEnum {
+    #[serde(rename = "CHILD")]
+    Child,
+}
