+++ src/lambda/src/format/detection.rs
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+
+#[derive(Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct Detection {
+    confidence: f64,
+    geometry: Geometry,
+    text: String,
+}
