+++ src/lambda/src/format/geometry/bouding_box.rs
@@ -1,0 +1,14 @@
+
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+
+#[derive(Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct BoundingBox {
+    height: f64,
+    left: f64,
+    top: f64,
+    width: f64,
+}
