+++ src/lambda/src/format/geometry/polygon.rs
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+
+#[derive(Serialize, Deserialize)]
+#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
+pub struct Polygon {
+    x: f64,
+    y: f64,
+}
