+++ src/lambda/src/format/input.rs
@@ -1,3 +1,7 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
 
 #[derive(Serialize, Deserialize)]
 #[serde(rename_all = "PascalCase")]
