+++ src/lambda/src/format/line_item/group.rs
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+
+#[derive(Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct LineItemGroup {
+    #[serde(rename = "LineItemGroupIndex")]
+    line_item_group_index: i64,
+    line_items: Vec<LineItem>,
+}
