+++ src/lambda/src/format/line_item/line_item_response.rs
@@ -1,0 +1,10 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+
+#[derive(Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct StateResponseLineItem {
+    line_item_expense_fields: Vec<StateResponseLineItemExpenseField>,
+}
