+++ src/lambda/src/format/summary/field_wrapper.rs
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+
+#[derive(Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+struct SummaryFieldsWrapper {
+    summary_fields: Vec<SummaryField>,
+}
+
+
