+++ src/lambda/src/format/summary/group_property.rs
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+
+#[derive(Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct GroupProperty {
+    id: String,
+    types: Vec<String>,
+}
