+++ src/lambda/src/main.rs
@@ -1,0 +1,21 @@
+use lambda_http::{run, service_fn, tracing, Body, Error, Request, RequestExt, Response};
+
+async fn function_handler(event: Request) -> Result<Response<Body>, Error> {
+    // get jsons
+    // format
+    // merge
+
+    let resp = Response::builder()
+        .status(200)
+        .header("content-type", "text/html")
+        .body(message.into())
+        .map_err(Box::new)?;
+    Ok(resp)
+}
+
+#[tokio::main]
+async fn main() -> Result<(), Error> {
+    tracing::init_default_subscriber();
+
+    run(service_fn(function_handler)).await
+}
